+++ lib/ui/home/HomeLaundryWidget.dart
@@ -240,7 +240,7 @@ class _HomeLaundryWidgetState extends State<HomeLaundryWidget> implements Notifi
     double? minContentHeight;
     for(GlobalKey contentKey in _contentKeys.values) {
       final RenderObject? renderBox = contentKey.currentContext?.findRenderObject();
-      if ((renderBox is RenderBox) && ((minContentHeight == null) || (renderBox.size.height < minContentHeight))) {
+      if ((renderBox is RenderBox) && renderBox.hasSize && ((minContentHeight == null) || (renderBox.size.height < minContentHeight))) {
         minContentHeight = renderBox.size.height;
       }
     }