+++ lib/ui/profile/ProfileHomePanel.dart
@@ -172,14 +172,16 @@ class _ProfileHomePanelState extends State<ProfileHomePanel> implements Notifica
           SingleChildScrollView(physics: _contentValuesVisible ? NeverScrollableScrollPhysics() : null, child:
             Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
               Padding(key: _pageHeadingKey, padding: EdgeInsets.only(left: 16, top: 16, right: 16), child:
-                RibbonButton(
-                  textStyle: Styles().textStyles.getTextStyle("widget.button.title.medium.fat.secondary"),
-                  backgroundColor: Styles().colors.white,
-                  borderRadius: BorderRadius.all(Radius.circular(5)),
-                  border: Border.all(color: Styles().colors.surfaceAccent, width: 1),
-                  rightIconKey: (_contentValuesVisible ? 'chevron-up' : 'chevron-down'),
-                  label: _getContentItemName(_selectedContent) ?? '',
-                  onTap: _onTapContentSwitch
+                Semantics(hint: Localization().getStringEx("dropdown.hint", "DropDown"), focused: true, container: true, child:
+                  RibbonButton(
+                    textStyle: Styles().textStyles.getTextStyle("widget.button.title.medium.fat.secondary"),
+                    backgroundColor: Styles().colors.white,
+                    borderRadius: BorderRadius.all(Radius.circular(5)),
+                    border: Border.all(color: Styles().colors.surfaceAccent, width: 1),
+                    rightIconKey: (_contentValuesVisible ? 'chevron-up' : 'chevron-down'),
+                    label: _getContentItemName(_selectedContent) ?? '',
+                    onTap: _onTapContentSwitch
+                  )
                 )
               ),
               _buildContent(),
@@ -213,8 +215,10 @@ class _ProfileHomePanelState extends State<ProfileHomePanel> implements Notifica
   Widget _buildContentDismissLayer() {
     return Positioned.fill(child:
       BlockSemantics(child:
-        GestureDetector(onTap: _onTapDismissLayer, child:
-          Container(color: Styles().colors.blackTransparent06)
+        Semantics(excludeSemantics: true, child:
+          GestureDetector(onTap: _onTapDismissLayer, child:
+            Container(color: Styles().colors.blackTransparent06)
+          )
         )
       )
     );
@@ -279,10 +283,10 @@ class _ProfileHomePanelState extends State<ProfileHomePanel> implements Notifica
 
   double? get _contentHeight  {
     RenderObject? pageRenderBox = _pageKey.currentContext?.findRenderObject();
-    double? pageHeight = (pageRenderBox is RenderBox) ? pageRenderBox.size.height : null;
+    double? pageHeight = ((pageRenderBox is RenderBox) && pageRenderBox.hasSize) ? pageRenderBox.size.height : null;
 
     RenderObject? pageHeaderRenderBox = _pageHeadingKey.currentContext?.findRenderObject();
-    double? pageHeaderHeight = (pageHeaderRenderBox is RenderBox) ? pageHeaderRenderBox.size.height : null;
+    double? pageHeaderHeight = ((pageHeaderRenderBox is RenderBox) && pageHeaderRenderBox.hasSize) ? pageHeaderRenderBox.size.height : null;
 
     return ((pageHeight != null) && (pageHeaderHeight != null)) ? (pageHeight - pageHeaderHeight) : null;
   }